Fix available water calculation in WithdrawGroundwaterIrrigation
The code for calculating per-layer water availability was inconsistent
between CalcAvailableUnconfinedAquifer and
WithdrawGroundwaterIrrigation. It seemed incorrect that, in
WithdrawGroundwaterIrrigation, irrig_layer always had zi - zwt, instead
of sometimes having zi(c,j) - zi(c,j-1) as in
CalcAvailableUnconfinedAquifer.

I think this problem is related to the more general bug in the
calculation of rsub_top_layer that @swensosc initially fixed in
ESCOMP/ctsm#523 but I asked him to back out of that PR (deferring its
fix to its own PR). Based on the fix he proposed there, it seems that
the code in CalcAvailableUnconfinedAquifer is correct, and the code in
WithdrawGroundwaterIrrigation is incorrect.

This commit fixes the issue.

Resolves ESCOMP/ctsm#595