Use aquifer_water_baseline*ratio rather than just aquifer_water_baseline
We can't reset wa_col to aquifer_water_baseline, or subtract
aquifer_water_baseline from wa_col, for tracers: instead, we need to use
the correct value for this tracer.

The resetting to aquifer_water_baseline at the start of BalanceCheck is
still problematic... we'll need to deal with that long-term.