Rework determination of available groundwater for irrigation
Rather than first computing available groundwater then later removing
it: instead, merge the functionality of CalcAvailableUnconfinedAquifer
and CalcIrrigWithdrawals into a single routine that accepts groundwater
irrigation demand and determines how to divide this demand into
layer-by-layer withdrawals from the unconfined aquifer as well as
withdrawal from the confined aquifer.

One motivation for this is to remove the duplication between
CalcAvailableUnconfinedAquifer and CalcIrrigWithdrawals. In addition,
this simplifies the code.

A note about the CalcIrrigWithdrawals interface: I explicitly pass in
all output arguments to CalcIrrigWithdrawals rather than passing the
whole waterfluxbulk_inst in order to make the outputs from this routine
explicit from the interface.