Exercise use_groundwater_irrigation block for most unit tests
I want to be sure we're testing various cases with
use_groundwater_irrigation True: especially cases where there isn't any
irrigation. This is so that we can exercise the code in this code block
in more circumstances.

So now I just turn off use_groundwater_irrigation in select tests where
this is intentional. (So note that there are still a few tests that
exercise the 'else' block.)