Fix WaterType unit tests for updated consistency check
Before this fix, two tests were failing because the bulk differed from
all tracers, whereas the assertExceptionRaised was only expecting one
failure. I tried to tell pFUnit to expect multiple failures, but
couldn't find a way to do that. The easiest solution seemed to be
changing the tests so that there would only be a single failure, so
that's what I have done here.