Fix job_launcher_no_batch unit test
Because subprocess.Popen starts the process in the background, it wasn't
finishing in time for the assertion in test_runCommand. Thus, the
assertion was failing. Adding a 'wait' fixes this.

Note that I made wait_for_last_process_to_complete work even if there
was no "last process", so that I can call it from the
test_runCommand_dryRun test, too.