Replace call to configure with a function that writes config_cache.xml
configure is now gone, so the build-namelist unit tests were failing
when they tried to call this now-nonexistent script. For clm45/clm50, in
an actual case, the python buildnml creates a config_cache.xml file. So
we replicate that functionality in build-namelist_test.pl.

In making this change, I changed most of the tests that did not specify
a phys version (and so I think were using clm4_0) to use
clm5_0. However, I removed a few clm4_0 tests that are no longer
applicable or are duplicated for clm4_5/clm5_0.