Revert change to use_init_interp defaults
This reverts part of 42b2b4991: I thought it was correct to change
use_init_interp to default to .false., now, but apparently not: with
that change, use_init_interp was staying .false. even when it needed to
be .true. (e.g., for a case with f10_f10_musgs grid).