Changed two new IHist tests to include decStart
Both are in the standard cheyenne intel testing suite.
One was introduced in this Pull Request (PR) #588.
The other was introduced in recent PR #483.

Both test the collapsing to fewer pfts than read from the input data
so decStart tests the year-end transition when pfts get updated in
transient pft (IHist) cases.