Change documentation of icemask_grc
The previous documentation (area in which GLC model can accept surface
mass balance) was true, but that wasn't the main meaning as far as CLM
is concerned: that meaning is more closely tied to
icemask_coupled_fluxes_grc. The more central meaning in CLM is: Area in
which values from GLC are valid.