Change the ciso bombspike tests to a single, global test
The single-point tests took a long time to run, and often timed out due
to system issues. In consultation with Erik, we felt that the important
thing was crossing the year boundary with this testing of c13 / c14 time
series, and that it's not too important to have tests both with and
without crop. So I am using a single, global test to cover this
requirement.

Ideally this test would include crops (in order to cover as much code as
possible combined with the c13/c14 timeseries, even though there are no
direct interactions between these timeseries and crops), but crop
DecStart tests currently fail because of
https://github.com/ESCOMP/ctsm/issues/404 and I didn't want to add
another long test just to test these options, so for now using a compset
without crops. Using a compset with SGLC to avoid problems with CISM in
DecStart tests; the only IHistClm50Bgc compset we have with SGLC is this
Qian compset, so I'm using this one.