Do not permit transient runs with n_dom_* > 0 or toosmall_* > 0
- These namelist parameters were introduced with NWP
(Numerical Wx Prediction) applications in mind, so transient
simulations are not an immediate priority.
- I placed error checks in CLMBuildNamelist.pm that follow how other
namelist options are handled:  if namelist options must not coexist
with do_transient_pfts and do_transient_crops, then the latter two get
switched to false and the simulation proceeds.
- I placed an error check that aborts in dynSubgridControlMod.F90 if
do_transient_pfts and do_transient_crops are still found to coexist
with n_dom_* > 0 and toosmall_* > 0