Update README files of two tests affected by the recent commits
Both tests are listed as IHist tests. Currently CLMBuildNamelist.pm resets
do_transient_pfts and do_transient_crops to .false. because it finds
n_dom_* > 0 or toosmall_* > 0. Resetting the do_transient_* varialbes allows
the tests to PASS. In the future we plan to allow coexistence of the
do_transient_* with the n_dom_* and toosmall_* variables.