Subtract virtual states to reduce dynbal fluxes for transient glaciers
Up until now, when computing the dynbal correction fluxes for transient
glacier columns, we have been (1) counting the mass and energy in the
ice column as if that is a real state, but on the other hand (2) NOT
accounting for the fact that glacier columns don't represent the soil
under glacier. These two issues work in opposite directions, but (1)
dominates, because there is much more ice (and energy, I think) in the
roughly 50 m of glacial ice than there is in a typical 50 m soil
column.

In discussing this issue with Bill Lipscomb, we came up with the idea of
subtracting some baseline value from each glacier column. I think that,
as long as we subtract the same baseline value throughout an entire
simulation for a given column, we will still conserve mass and energy
through dynamic landunit transitions.

So, here we subtract baseline values from glacier columns, accounting
for the two issues mentioned above: (1) we subtract the water and energy
in the glacier ice, because this is a virtual state in CTSM, and (2) we
add the water and energy from the vegetated column(s), to account for
the fact that we don't have an explicit representation of
soil-under-glacier (this carries the assumption that the
soil-under-glacier has the same state as the initial vegetated state in
that grid cell). We currently set these baselines in initialization, so
they are set equal to the cold start state. Water and ice in the glacial
ice stay fixed over the course of a simulation, so the cold start values
should be the same as the current values at any point in time. The heat
content of the glacial ice does change over time, but by subtracting
this baseline value, we should be able to reduce the dynbal sensible
heat fluxes.

In a following commit, I plan to allow the user to reset these baselines
at some desired point in the simulation. I think that, in general, this
resetting would break conservation. But as long as it is done before the
onset of transient glaciers, I think this should be okay. In this way,
the user can minimize dynbal fluxes even further, by resetting the
baselines after the system has spun up.

Partially addresses ESCOMP/ctsm#274