Do not subtract dynbal baselines from begwb and endwb
Ideally, we would keep begwb and endwb consistent with liq1, liq2, ice1
and ice2 in this respect - i.e., subtracting the dynbal baselines in all
cases. However, this changes answers for methane in a way we want to
avoid right now.

See https://github.com/ESCOMP/ctsm/issues/659 for details.