Use liqcan and snocan in place of h2ocan
This is the first step towards making h2ocan just a diagnostic quantity
that is just computed once near the end of the time step, and not used
anywhere in the prognostic equations. This commit introduces a minimal
set of changes to replace h2ocan with liqcan and snocan. I expect these
diffs to introduce roundoff-level answer changes. I have put in place
temporary code to ensure that changes are no greater than roundoff
level.

There is a lot of mess at this point (bad variable names, unnecessary
code, etc.). I'll clean this up in a following bit-for-bit commit.