Remove check that h2ocan > 0 iff h2ocan_new > 0
This was triggering an error. The one example of the error that I looked
at was:

 FracWet: h2ocan > 0 but not h2ocan_new:
          13  3.067113237684536E-023 -1.734692804844432E-018
 -1.734723475976809E-018  3.067113237684536E-023

From plugging in some example values, it looks like, as long as we
ensure that h2ocan and h2ocan_new differ by no more than 1e-13, the
difference due to entering vs. not entering this conditional will be
relatively small for fwet (fwet up to about 1e-8 rather than 0). The
difference is more significant for fcansno (fcansno up to about 0.01
rather than 0). But for now I'm going to treat this as acceptable. If
the test suite passes with this, I'll check with others if this seems
okay - i.e., if it seems okay that we sometimes aren't entering this
code block now whereas we were before, or vice versa, when h2ocan or
h2ocan_new are roundoff-level different from 0.