Rename snowveg_flag to snowveg_affects_radiation; always compute fcansno
Previously, this flag affected behavior in CanopyHydrologyMod and
SurfaceAlbedoMod. However, the effects in CanopyHydrologyMod were
unnecessary: We can simplify things and make it a bit easier to
understand the code by always computing fcansno, even if it isn't
currently needed. (This has a minor performance cost, but since we
typically have snowveg_affects_radiation on anyway, it seems worth doing
this cleanup.)

This namelist flag is now private to SurfaceAlbedoMod.