Turn off MEGAN with NWP
This is wanted for performance reasons.

I debated doing this via the nwp configuration or via the -no-megan flag
that could be set by the compset. I opted for the former because it will
be easier to remember what differs between an nwp vs. clm
configuration (it's easier to remember that two things differ
(configuration and structure) than three (configuration, structure and
the -no-megan flag).