Change year-2000 compsets to start in year 2000; similar for 2010
This was requested by Sean Swenson and Dave Lawrence, so that the model
year for year-2000 compsets will match real years rather than arbitrary
years starting with 1.

There are no year-2010 compsets on master, but there is one on the
release branch, so the change to year-2010 compsets will take effect
once the release branch is merged back to master.