Remove excess canopy liquid/snow regardless of temperature
Previously: if it was above freezing, snow was allowed to accumulate
beyond the specified capacity, and if it was below freezing, liquid was
allowed to accumulate beyond the specified capacity. Sean Swenson, Keith
Oleson and Dave Lawrence agreed that this seems wrong.

This commit changes the behavior so that both liquid and snow canopy
capacities are enforced regardless of temperature.

Resolves ESCOMP/ctsm#699