First stage of major CanopyHydrology refactor to support water tracers
This commit breaks out the first section of CanopyHydrology (the part
that is actually doing canopy hydrology) into a bunch of smaller
routines in order to support the addition of water tracers. The
important consideration here is that flux calculations need to be
cleanly separated from state updates so that, in between each flux
calculation and state update, we can set the relevant tracer fluxes.

See also
https://github.com/billsacks/prototypes-ctsm-canopyhydrology_tracers