Add more temporary code to get bit-for-bit
On master, the max statements are only executed if we're within the
conditional (which here is given by
check_point_for_interception_and_excess). It seems simpler and possibly
numerically preferable to always do these max statements, but for now
I'm putting in some code to ensure bit-for-bit with master.