Make wetland snow resetting behavior more explicit
It seems like the intended behavior here is to zero out the entire snow
pack. Currently, however, this only zeros out a very thin (zero-layer)
snow pack. For now, I'm adding an snl==0 conditional to make this
behavior explicit; long-term, we'd like to change this to actually zero
out the whole snow pack. (At that time, this code block should probably
be moved to a more appropriate home, as noted in comments in issue
ESCOMP/ctsm#735.)