Recalculate h2osno for the sake of SnowCapping
SnowCapping decides whether to remove excess snow based on
h2osno. However, at the point where this is called,h2osno is out of sync
with (h2osoi_liq + h2osoi_ice): Those variables can be changed slightly
in SnowWater (which is called before SnowCapping), yet h2osno isn't
recalculated based on those changes.

I don't think this causes a big problem in practice, but it seems like
it would improve code understandability if we had the general rule that
any code that references h2osno is using a version of h2osno that is in
sync with (h2osoi_liq + h2osoi_ice). This is also needed to prevent
answer changes in an upcoming major refactor.

I have confirmed that this change will change answers, by running a
one-year test with and without the change in this
commit (SMS_Ly1.f09_g17.I2000Clm50SpGs.cheyenne_intel.clm-monthly). Answer
changes are expected to be small, since this should just change the
timing of snow capping a bit, but I haven't looked closely at the
magnitude of the answer changes.

Resolves ESCOMP/ctsm#736