Remove temporary code to ensure only roundoff-level diffs
I have ensured that, with this code in place, diffs are only
roundoff-level for most tests (there were a few exceptions, where the
number of snow layers differed on the initial conditions file as
compared with the current run; I have separately verified what's going
on in those).

So now I can remove this temporary code.