Add col%lun_itype
col%lun_itype(ci) is the same as lun%itype(col%landunit(ci)), but is
often a more convenient way to access this type

I got tired of having to get the landunit index just for the sake of
getting the landunit's type of a given column.

I have NOT gone through the code and replaced lun%itype(l) with
col%lun_itype(c) where applicable, but this will be useful for the
future.