First steps toward tracerization of FracH2oSfc, and change filter
Put a wrapper around FracH2oSfc (now BulkDiag_FracH2oSfc) as the first
step towards tracerization. Also, change the filter for this routine
from nolakec to soilc, since there was previously a conditional that
ended up accomplishing the same thing. (I have confirmed that
frac_h2osfc is initialized to 0 everywhere, so the explicit setting to 0
for special landunits should be unnecessary.)