Change logic for rain-snow conversion terms
Use rain diff in rain to snow, snow diff in snow to rain.

Rationale: this is less sensitive to roundoff errors in the case where
we are bringing rain exactly to 0, but rain << snow (or vice
versa). This should allow us to return to using a tolerance of 1e-13 in
AdjustPrecipTracer. Even with 1e-12, I was getting a failure in the
tracer consistency check on hobart_nag:

rain_orig = 0.4384091435227974479491E-09
snow_orig = 0.1401496725888953037915E-04
snow (new) = 0.1401540566803305240442E-04
rain_to_snow = 0.4384091435220252748473E-09

Since rain was about 5 orders of magnitude less than snow, and we
compute rain to snow conversion as:

    rain_to_snow_conversion_c(c) = forc_snow_c(c) - snow_orig

we ended up introducing a ~ 1e-11 relative error in rain to snow
conversion. We could have arbitrarily high errors with the old formula.