Remove duplicated code by using new function from SnowCompaction
There was a minor difference between the formulation in SnowCompaction
and the one in FracSnowDuringMeltClm5: the latter has an extra min
statement. It looks like this shouldn't change answers (the min
statement shouldn't be necessary, given the min that is already applied
to smr), but it's possible that it will change answers by roundoff in
rare cases.