Use longer names for parameterization names
Sean Swenson and others suggested using these longer names rather than
the more cryptic short abbreviations. Part of the motivation is once we
have 4 separate files, this will make the science modules stand out more
because of their longer names.

I wanted to name the namelist group scf_SwensonLawrence2012_inparm for
consistency with the namelist option, but this didn't work with the
namelist generation scripts, presumably because of the capitalization;
so I'm using the all-lowercase form, scf_swenson_lawrence_2012_inparm.