Allow snow cover fraction types to have different init interfaces
Since the init method is only called from the factory method, it's
fairly straightforward to allow the different init methods to have
different interfaces, rather than having a bunch of unused arguments in
some of the methods. The big motivation here is avoiding the need to
change all parameterizations' Init methods if you add a new
parameterization with a new input requirement.

This ended up being a little non-trivial because of the need for the
select type statement. I think we could have avoided that if we did all
initialization in the constructor (avoiding an Init method
entirely). But we have had some compilers (e.g., intel17) that don't
seem to work right with function constructors, so for now we're sticking
with an Init subroutine.