Always set frac_sno_eff to 1 for lakes
This is not expected to change answers, since frac_sno_eff should be
ignored when there is no snow. Indeed,
SMS_D_Ld1_P4x1.f10_f10_musgs.I2000Clm50BgcCropQianRsGs.bishorn_gnu.clm-default
is bit-for-bit except for FSNO_EFF itself.

The motivation for this is: We want to share more code between lake and
non-lake subroutines; this is a step in that direction.