Remove qflx_snow_grnd > 0 criteria for lake snowpack initiation
This criteria is not used for other landunits, so remove it for lakes in
order to reconcile the two. In principle this could change answers. My
guess, though, is that this won't change answers, because probably
qflx_snow_grnd must be > 0 if we have both snl(c) == 0 and snow_depth >
threshold - but I'm not sure about this.

So far I have just run the test
SMS_D_Ld1_P4x1.f10_f10_musgs.I2000Clm50BgcCropQianRsGs.bishorn_gnu.clm-default,
which does NOT change answers.