For lakes, also check qflx_snow_grnd for snowpack initiation
I'm not sure why this is needed for lakes when it isn't done for other
landunits. However, I was getting answer changes relative to master, and
I suspect my dropping of this condition was responsible. So I'm putting
this in to try to get bit-for-bit answers with master.

I also refactored the inclusion of an additional depth: Since I now have
a lake conditional in BuildFilter_SnowpackInitialized anyway, I am
directly referencing lsadz there rather than awkwardly passing it down.