Remove near-duplicated code setting qflx_sl_top_soil
For landunits other than soil, crop and urban, this new version isn't
exactly the same as the old version, but I think it's still correct. In
particular: If j == 0: now we always set this; before, it was only set
for these special landunits if snl(c) < -1... i.e., now we also set it
if there is a single snow layer. I don't think this will change any
behavior, because in this case there was a single snow layer that is in
the midst of disappearing, so this column would no longer have snl(c) <
0, and so the value of this flux would be irrelevant (since it is only
referenced if snl(c) < 0 by the time of the balance check). It's simpler
just to always set it, as is done for veg. and urban landunits, so
that's what I've done in the refactor.