When snow layers disappear, don't bother setting h2osoi_liq_col(c,0) = 0
My original concern was that this was being done for everything except
lake, which seemed weird. But upon further reflection, this seems
unnecessary, since this should be done soon afterwards in
ZeroEmptySnowLayers.

I'm not 100% sure this removal is safe / correct. I have tested it in
SMS_D_Ld1_P4x1.f10_f10_musgs.I2000Clm50BgcCropQianRsGs.bishorn_gnu.clm-default,
which is still bit-for-bit. I will do more extensive testing on this to
ensure that this change doesn't change answers.