Tracerize ZeroEmptySnowLayers
As for CombineSnowLayers and DivideSnowLayers, I did this by inserting
tracer updates in the middle of the existing loop, rather than splitting
out multiple routines. This felt cleaner, particularly because the
latter approach would require duplication of the logic determining which
layers should be zeroed; also, this routine feels vaguely similar to
CombineSnowLayers and DivideSnowLayers, so I felt justified in taking
this approach because it's the same as the approach I took for those two
routines.

This completes the tracerization of the bulk of the snow code - most or
all of the code in SnowHydrologyMod.