Remove eflx_grnd_lake restart field
It looks like this is no longer needed. This was added to the restart
file in clm4_5_1_r105 (cf8a5e0), presumably to fix the different fill
patterns upon restart that are mentioned in ESCOMP/ctsm#31. However,
there was a later change that seems to fix this issue the correct way,
by ensuring that eflx_grnd_lake is set in every time step, rather than
potentially persisting at its previous value (in clm4_5_13_r204 -
ce9e4075; see also the discussion in
http://bugs.cgd.ucar.edu/show_bug.cgi?id=1717).

Note that eflx_grnd_lake is just a diagnostic field: it doesn't appear
on the right-hand side of any equations in the model.