BTRAN diagnostic fields: exclude special landunits from averages
Previously, BTRAN and BTRANMN were being output as 0 for special
landunits, potentially contaminating grid cell averages. This commit
ensures that BTRAN, BTRANMN and BTRAN2 are only averaged over vegetated
landunits. I think this was already the case for BTRAN2, but I'm making
the change for that field, too, for consistency.

Resolves ESCOMP/ctsm#27