Change 'debug' back to parameter, change its name
Mariana said that there were problems with having this be a parameter
named debug (something related to conflicts with the DEBUG cpp token),
so she had changed it to not be a parameter. But I think the correct fix
here is to change its name to avoid conflicts with the DEBUG cpp token,
and then change it back to being a parameter.