Handle issue_fixed metadata properly on interpolation
The previous implementation didn't work right when doing an interpolated
start: The new issue-fixed global attribute was being written to the
finidat_interp_dest file, making it seem like the information on that
file had already been fixed. The correct behavior is that this attribute
should be taken from the source file used in interpolation: if the issue
was fixed on the source file, then it is fixed on the destination file,
too; otherwise, it isn't.

This commit introduces that behavior. In addition, because the logic for
dealing with this issue-fixed metadata is getting more complex, I have
made a new module to deal with this.