Additional changes that were required for mkpft to work
In testing for the last commit, I had forgotten to change map_fpft in my
namelist to point to the nomask instead of the modis version of this
file. As soon as I changed the namelist I got an error that one of the
pft variables was not summing to 100. This commit includes the necessary
changes for the code to work again.