Correct arguments to setup_logic_soilm_streams
Properly send $opts down to it, and use it in calls inside.
Also add a check that this isn't a transient use case for soil
streams with linear time-interpolation as that will likely cause
an error going from a vegetated point that is missing to not, or
vice versa. Also add a check for that. And fix the soilm test which
had phys in place of conopts.