Add tintalgo for: lai, lightning, pop-dens, urbantv, and ndep
Add tintalgo settings for most of the streams. Also
change the buildnml fatal-error for soilm_tintalgo='linear'
for transient to a warning that can be overridden if desired.
Also fix the build-namelist unit test for warnings, so it
verifies that warnings die, but will run if -ignore_warnings
is set. Also get maxpatch_pft to normally die if the two
valid settings are used (79 or 17).