Fix ctsm.mk so that share code is included in the include paths
This is needed for compilation to work. Note that my last 3 commits
(where I was messing with the build) were originally on top of
63beebb3, then I rebased them to be up to date with the lilac_cap
branch. When I did this rebase, the build started failing; this fixes
it.

I thought that we could skip SHR_INC in CTSM_INCLUDES: this include is
needed when compiling CTSM source, but I didn't think it would be
needed when compiling the atmosphere. However, without that, I get an
error when compiling atm_driver.F90:

/glade/work/sacks/ctsm_code/current_branch1/lilac/atm_driver/atm_driver.F90(23):
error #7002: Error in opening the compiled module file.  Check INCLUDE
paths.   [SHR_KIND_MOD]
  use lilac_atmcap, only : lilac_atmcap_atm2lnd, lilac_atmcap_lnd2atm
------^

So apparently it needs to be able to find SHR_KIND_MOD due to its
indirect use? (There are also errors due to new direct uses of share
code from atm_driver.F90, but the surprising thing to me was this
apparent need to have indirectly-used modules in the include path.)