Remove some unused code from the time manager
The point of this removal is to facilitate removing stop time from the
time manager. So I have removed some unused routines related to stop
time. I have also removed the unused get_clock because I want to prevent
anyone from trying to get the clock and then trying to get the
soon-to-be-non-existent stop time out of the clock.