Don't set a real stop time on CTSM's internal clock
Stop time was being set but never used. For LILAC, it is awkward to
require the stop time to be set in initialization. So I'm removing the
need for it entirely. We instead set a dummy stop time far in the
future. (If we were using the real ESMF time manager, we could avoid
setting stopTime on the clock, but it appears that cime's ESMF time
manager requires stopTime to be set.)

Also remove unused get_timemgr_defaults.