Remove unused lat_rounded and lon_rounded
These were originally here for the sake of comparison against datm: The
idea was that lats and lons could differ by a little between datm and
this dummy driver, but the rounded-to-the-nearest-integer values should
be the same. But it looks like either these were never used, or they
were used at one point but then stopped being used.