Write fields sent to atm at end of run
This is useful for validating the fields sent from CTSM to the
atmosphere, ensuring that that side of the data flow is working
right. I'll use it to do a one-time manual validation, then it will also
be leveraged in the future to ensure that this side of the data flow
doesn't break, by checking these files in our automated testing of lilac
with the demo atm driver.