Check userRc in addition to rc
Some ESMF routines have an optional userRc argument that holds the
return code from the user callback that is called from these ESMF
routines. It is important to check this argument, too, in order to catch
errors that occur in the user routine. In fact, most errors will show up
in this argument rather than the standard rc argument: the userRc
argument captures any error that is raised in routines like lnd_init and
lnd_run.