Update cmeps version
I started getting errors in preview_namelists with the old
version:

Traceback (most recent call last):
  File "/Users/sacks/ctsm/ctsm2/cime/src/drivers/nuopc/cime_config/buildnml", line 509, in <module>
    _main_func()
  File "/Users/sacks/ctsm/ctsm2/cime/src/drivers/nuopc/cime_config/buildnml", line 506, in _main_func
    buildnml(case, caseroot, "drv")
  File "/Users/sacks/ctsm/ctsm2/cime/src/drivers/nuopc/cime_config/buildnml", line 474, in buildnml
    _create_drv_namelists(case, infile, confdir, nmlgen, files)
  File "/Users/sacks/ctsm/ctsm2/cime/src/drivers/nuopc/cime_config/buildnml", line 241, in _create_drv_namelists
    _create_runseq(case, coupling_times)
  File "/Users/sacks/ctsm/ctsm2/cime/src/drivers/nuopc/cime_config/buildnml", line 332, in _create_runseq
    from runseq_I import runseq
  File "/Users/sacks/ctsm/ctsm2/cime/src/drivers/nuopc/cime_config/../../../../src/drivers/nuopc/cime_config/runseq/runseq_I.py", line 70
    print "cism_evolve = ",cism_evolve
                         ^
SyntaxError: Missing parentheses in call to 'print'. Did you mean print("cism_evolve = ",cism_evolve)?
ERROR: /Users/sacks/ctsm/ctsm2/cime/src/drivers/nuopc/cime_config/buildnml /Users/sacks/projects/scratch/test_lilac_0309b FAILED, see above

The immediate error was a lack of parentheses in the print function
(needed with python3), but the weirder error is that this is in a code
block that should only be executed if using cism (not stub glc), but
this was in a compset using stub glc. I'm not sure why this started now,
but updating to the latest cmeps version seems to fix the issue, so I'm
doing that.