Only set Sa_topo in the first time step
This is so that we test the infrastructure that allows fields to be just
set once. (To some extent that was already tested via Sa_landfrac, but
this is a field that is actually accessed each time step inside CTSM, so
this provides a better test that we're actually sending the correct
field across each time step.)

This is bit-for-bit with the last commit, as desired.