Update initial conditions file used in LII test
Without this change,
LII_D_Ld3.f19_g17_gl4.I2000Clm50BgcCrop.cheyenne_intel.clm-glcMEC_spunup_1way
failed because it's pointing to an old initial conditions file that has
a certain variable as a double when it is expected to be an int. I guess
this worked with the old ncdio_pio code but breaks with the ncdio_pio
changes on this branch. I have updated the initial conditions file using
the finidat_inter_dest.nc file from this test, as described in the
README file.