Remove near-duplicated Makefiles and conf.py files
Previously, we had been (sort of) supporting builds of just the User's
Guide or just the Tech Note via near-duplicated copies of the Makefile
and conf.py file. However, these didn't really seem to work right for an
html build (the files ended up in the wrong directory structure), and
this was going to be harder to support and maintain moving forward,
because of the duplication. Erik Kluzek and Keith Oleson have confirmed
that they don't actually need support for an html build of just part of
the documentation. I have made changes to the main Makefile to support
just building a pdf of the tech note (not including the User's Guide),
which IS needed.