Fix bug in snow aerosol numerics
Multiplications / divisions by dtime were missing in a few places. This
bug was introduced in ctsm1.0.dev065, when a term was changed from a
mass to a flux (per unit time), but some of the code wasn't changed
correspondingly.

It appears that this bug only appears infrequently, but when it does, it
can cause snow aerosol concentrations to blow up, becoming about 20
orders of magnitude too large.

Resolves ESCOMP/ctsm#965