Changes needed so that fates_fire_data_method is local to clmfates_interface
These are the changes that I see that are needed  to keep fates_fire_data_method
local to clmfates_interface. Whenever fates_fire_data is accessed it has to be
done through clmfates_interface, because that's the object that contains it and has
it instantiated. We also will want it to be private to clmfates_interface, so it
can't be accessed or changed outside of there.

There were some stubs here to pass specific instances of the class types before, but
they weren't completed. So those are all removed and the generic method is being
accessed so that you don't have to know which specific instance is being used.

Testing: I haven't tested it yet. But, will before pushing