Make lnfm24 data private add an get method for it
Make the lnfm24 data private to inside of FATES_FireData object
and add an accessor for it. The accessor gets the pointer, so there's
no time spent on copying although using a pointer this way also means
you can screw up  the  data outside of the object. But, it seems
fairly well  constrained so I don't know that's a problem.

This is still not compiling  because there needs to be a FATES
interface to the object that has the  FATES specific new methods
that  have been added.