Change smallville ERS_Lm20 test to be cold start
This was the intent when I first created the test: The point of this
test is to test crop restart shortly after cold start. It later
accidentally turned into a test that used spun-up initial conditions
(when we changed just about everything to use spun-up initial
conditions); this commit restores it to being cold start.