Specify path to commands
Now, for commands executed from the case directory, we specify the path
to the case directory both in the command itself and in the cwd
argument. We do the former in case dot isn't in the user's path; we do
the latter in case the commands require you to be in the case directory
when you execute them.