Oet rid of most of the -no-crop cases
Get rid of most of the no-crop cases since now we can use
crop files for no-crop as well. Also remove default creation
of T31 and hcru resolutions.

Keep -no-crop cases for f09,f19,f10 with vic and T42 for SCAM
and nldas for NWP. Setup the historical now runs SSP5-8.5 so
it can be used for both future, and historical for beyond 2015.

Fixes: #1025 By default only create crop version of surface/landuse.timeseries datasets