LILAC build: make default withOUT OpenMP support
I found that on cheyenne, if you don't explicitly specify
OMP_NUM_THREADS in your job, you get 36 threads per node. This is a huge
problem for performance, and is something that could easily happen to a
user who wasn't planning to run threaded anyway. So it feels like the
safest thing to do is to build without threading by default.