Use a stub rof rather than mosart with lilac
This change is important in the short-term, at least, because I'm not
ready to bring the mosart lilac_cap branch to master, so we're pointing
to a version of mosart that doesn't have the necessary changes - see
https://github.com/ESCOMP/MOSART/pull/32.

Once those MOSART changes are on MOSART's master branch, then we should
change buildlib back to using the MOSART source code rather than stub
rof. We may want to do this conditionally, depending on whether rof
coupling is actually wanted in the given run.

(I at first thought that we could let the cime build build mosart for
us, but then realized that the current mechanism is needed because lilac
depends on the mosart code; also, mosart is not built during the
--sharedlib build phase.)

Note: I have NOT given careful thought to the changes in lilac_mod.F90:
It seems right to put this rof-related code inside a conditional, but I
haven't done a careful analysis to determine if that's correct.