LILAC build_ctsm: Stop making sym link to existing inputdata directory
The initial motivation for this change was that the LILAC system test
was failing because the full path to the aerosol deposition file was too
long (longer than 256 characters), because the path to the inputdata sym
link was very long. I thought about allowing a longer name, but this
would involve changes in CIME's stream code, and I didn't want to go
there. So I decided to stop using this sym link and instead just use the
true path to the inputdata directory.

After making this change, I started feeling like this is an improvement
even without the motivation of the system test: sym links can be
confusing, and it seems more intuitive to just point to files in their
"real" location.