Fix data type of haslake variable
This is a double precision variable on the file, so I'm making it double
precision in the code. I think ncdio_pio used to handle data type
conversions like this, but now it seems to be corrupting memory: the
original code led to errors in decompInitMod, due to what appears to be
corruption in some unrelated variable.