Add new runtime consistency checks for do_transient_lakes
I'm treating do_transient_lakes similar to do_transient_pfts and
do_transient_crops in this respect. I'm not sure if all of these checks
are truly important for transient lakes (in particular, my guess is that
collapse_urban could probably be done with transient lakes - as well as
transient pfts and transient crops for that matter), but some of the
checks probably are needed, and it seems best to keep transient lakes
consistent with other transient areas in this respect.