do_transient_lakes in namelist: set default value and do error checks
The new function setup_logic_do_transient_lakes is based on
setup_logic_do_transient_crops.

As with my previous commit: I'm not sure if all of these checks are
truly important for transient lakes (in particular, my guess is that
collapse_urban could probably be done with transient lakes - as well as
transient pfts and transient crops for that matter), but some of the
checks probably are needed, and it seems best to keep transient lakes
consistent with other transient areas in this respect.