Reset lake dynbal baselines if using an old initial conditions file
We have changed the definition of total column water for lake columns,
so the baseline values for lakes are incorrect on old initial conditions
files. This commit adds some code to check if we're using an old initial
conditions file, and if so, resets the dynbal baseline values for lakes
to use the new definition.