Reset lake baselines in branch or continue run, too
I went back and forth about whether we should do this, but I actually
feel that it's best if we do reset the lake baselines in a branch or
continue run, if using an older restart file. If we didn't do this, we'd
want to add some logic for writing out the issue-fixed metadata for any
further restart files written from these runs, to note that this issue
isn't actually fixed yet on these restart files.